Playback working

+++ ffmpeg-sharp/src/Interop/AVCodec/AVPicture.cs
@@ -35,7 +35,7 @@ namespace FFmpegSharp.Interop.Codec
     [StructLayout(LayoutKind.Sequential)]
     public unsafe struct AVPicture
     {
-        public fixed byte data[8];
+        public fixed int data[8];
         /// <summary>
         /// number of bytes per line
         /// </summary>
